Quick patch to allow space between -p - Also remove compiler warning
svn-id: r4065

+++ script_v2.cpp
@@ -23,7 +23,9 @@
 
 #include "stdafx.h"
 #include "scumm.h"
+
 #include "sound/mididrv.h"
+
 #include "sound/imuse.h"
 
 void Scumm::setupOpcodes2()
@@ -2677,6 +2679,7 @@ void Scumm::o6_miscOps()
 	int16 args[30];
 	int i;
 	Actor *a;
+
 	IMuse *se = _imuse; //yazoo: not very nice
 
 	getStackList(args, sizeof(args) / sizeof(args[0]));
@@ -2849,7 +2852,8 @@ void Scumm::o6_miscOps()
 			break;
 
 		case 122:
-			_vars[VAR_SOUNDRESULT] = se->do_command(	args[1], args[2], args[3], args[4],
+
+			_vars[VAR_SOUNDRESULT] = (short)se->do_command(	args[1], args[2], args[3], args[4],
 							args[5], args[6], args[7], args[8]
 							);
 			break;
